+++ iphonePlay/Classes/LowLevelSoundGeneratorProtocol.h
@@ -6,12 +6,13 @@
  *  Copyright 2009 BestBefore Ltd. All rights reserved.
  *
  */
+#import <CoreGraphics/CoreGraphics.h>
 
 @protocol LowLevelSoundGeneratorProtocol
 
 - (int)nextFreeChanel; // will return -1 if no free channels
 
-- (void)turnOnSine:(UInt8)inIndex freq:(float)freq amp:(CGFloat)amplitude;
+- (void)turnOnSine:(UInt8)inIndex freq:(CGFloat)freq amp:(CGFloat)amplitude;
 - (void)turnOffSine:(UInt8)inIndex;
